28 add movie top lists features

+++ src/Via.Movies.Api/Repositories/IDirectorRepository.cs
@@ -4,7 +4,7 @@ namespace Via.Movies.Api.Repositories;
 
 public interface IDirectorRepository
 {
-	Task<Person?> GetDirectorOfMovie(int movieId);
+	Task<Person?> GetDirectorOfMovie(long movieId);
 	Task<IEnumerable<Director>> GetAllDirectorsAsync();
 	Task<Director> CreateDirectorAsync(Director director);
 }
